Move salt.utils.context tests to their proper location
